feature(fragment): extracted all "public types" into separate files
