+++ packages/groqd/src/commands/order.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect, expectTypeOf } from "vitest";
 import { SanitySchema, q, zod } from "../tests/schemas/nextjs-sanity-fe";
-import { InferResultType } from "../types/public-types";
 import { executeBuilder } from "../tests/mocks/executeQuery";
 import { mock } from "../tests/mocks/nextjs-sanity-fe-mocks";
+import { InferResultType } from "../groq-builder";
 
 const qVariants = q.star.filterByType("variant");
