+++ packages/groqd/src/commands/raw.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, expectTypeOf, it } from "vitest";
 import { q, zod } from "../tests/schemas/nextjs-sanity-fe";
-import { InferResultItem, InferResultType } from "../types/public-types";
 import { executeBuilder } from "../tests/mocks/executeQuery";
 import { mock } from "../tests/mocks/nextjs-sanity-fe-mocks";
+import { InferResultItem, InferResultType } from "../groq-builder";
 
 describe("raw", () => {
   const qRaw = q.star
