+++ packages/groqd/src/commands/root/value.ts
@@ -1,5 +1,5 @@
 import { GroqBuilderRoot } from "../../groq-builder";
-import { Parser } from "../../types/public-types";
+import { Parser } from "../../types/parser-types";
 
 declare module "../../groq-builder" {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
