+++ packages/groqd/src/commands/slice.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect, beforeAll, expectTypeOf } from "vitest";
 import { SanitySchema, q } from "../tests/schemas/nextjs-sanity-fe";
-import { InferResultType } from "../types/public-types";
 import { executeBuilder } from "../tests/mocks/executeQuery";
 import { mock } from "../tests/mocks/nextjs-sanity-fe-mocks";
+import { InferResultType } from "../groq-builder";
 
 describe("slice", () => {
   const qVariants = q.star.filterByType("variant");
