+++ packages/groqd/src/types/public-types.ts
@@ -1,170 +1,25 @@
-import type { ZodType } from "zod";
-import type { ResultItem } from "./result-types";
-import type { QueryConfig } from "./query-config";
+export { ExtractDocumentTypes, SchemaDocument } from "./document-types";
+export {
+  InferFragmentType,
+  FragmentResultTypeTag,
+  Fragment,
+} from "./fragment-types";
+export {
+  InferParserInput,
+  InferParserOutput,
+  Parser,
+  ParserFunction,
+  ParserObject,
+} from "./parser-types";
+export { ExtractRefType, RefType } from "./ref-types";
 
-/* eslint-disable @typescript-eslint/no-explicit-any */
-
-/**
- * A Parser is either a generic mapping function, or a Zod schema.
- * It's used for run-time validation and/or transformation
- * of the results of a field.
- */
-export type Parser<TInput = any, TOutput = any> =
-  | ParserFunction<TInput, TOutput>
-  | ZodType<TOutput, any, TInput>;
-
-/**
- * Same as `Parser`, except it allows for wider input types,
- * so that a value of `string` can be handled
- * by a parser that accepts `string | null` etc.
- */
-export type ParserWithWidenedInput<TInput> =
-  // TypeScript automatically widens the parameters of a function:
-  | ParserFunction<TInput, any>
-  // TypeScript doesn't widen types for the ZodType signature;
-  // (but we type-check this manually in the ExtractProjectionResult)
-  | ZodType<any, any, any>;
-
-export type InferParserInput<TParser extends Parser> = TParser extends Parser<
-  infer TInput
->
-  ? TInput
-  : never;
-export type InferParserOutput<TParser extends Parser> = TParser extends Parser<
-  any,
-  infer TOutput
->
-  ? TOutput
-  : never;
-
-/**
- * A generic "parser" object which can take any input and output a parsed type.
- */
-export type ParserObject<TInput = any, TOutput = any> = {
-  parse: ParserFunction<TInput, TOutput>;
-};
-
-/**
- * A generic "parser" function which takes any input and outputs a parsed type.
- */
-export type ParserFunction<TInput = any, TOutput = any> = (
-  input: TInput
-) => TOutput;
-
-export type ParserFunctionMaybe<
-  TInput = any,
-  TOutput = any
-> = null | ParserFunction<TInput, TOutput>;
-
-// Not used at runtime, just used to infer types:
-export declare const GroqBuilderResultType: unique symbol;
-export declare const GroqBuilderConfigType: unique symbol;
-
-/**
- * IGroqBuilder contains the minimum results of a GroqBuilder chain,
- * used to prevent circular references
- */
-export type IGroqBuilder<
-  TResult = unknown,
-  TQueryConfig extends QueryConfig = QueryConfig
-> = {
-  /**
-   * Used to infer the TResult types of a GroqBuilder.
-   * This symbol is not used at runtime.
-   * @internal
-   */
-  readonly [GroqBuilderResultType]: TResult;
-  /**
-   * Used to infer the TQueryConfig types of a GroqBuilder.
-   * This symbol is not used at runtime
-   * @internal
-   */
-  readonly [GroqBuilderConfigType]: TQueryConfig;
-  /**
-   * The GROQ query as a string
-   */
-  readonly query: string;
-  /**
-   * The parser function that should be used to parse result data
-   */
-  readonly parser: ParserFunction | null;
-  /**
-   * Parses and validates the query results, passing all data through the parsers.
-   */
-  readonly parse: ParserFunction;
-};
-
-export function isGroqBuilder(
-  fieldConfig: unknown
-): fieldConfig is IGroqBuilder {
-  return (
-    !!fieldConfig &&
-    typeof fieldConfig === "object" &&
-    "query" in fieldConfig &&
-    "parse" in fieldConfig &&
-    typeof fieldConfig.query === "string" &&
-    typeof fieldConfig.parse === "function"
-  );
-}
-
-/**
- * Represents a GroqBuilder chain that is "terminal",
- * and should not be further chained.
- */
-export type IGroqBuilderNotChainable<
-  TResult,
-  TQueryConfig extends QueryConfig
-> = IGroqBuilder<TResult, TQueryConfig>;
-
-/**
- * Extracts the Result type from a GroqBuilder query
- */
-export type InferResultType<TGroqBuilder extends IGroqBuilder<any>> =
-  TGroqBuilder extends IGroqBuilder<infer TResultType> ? TResultType : never;
-
-/**
- * Extracts the Result type for a single item from a GroqBuilder query
- */
-export type InferResultItem<TGroqBuilder extends IGroqBuilder<any>> =
-  ResultItem.Infer<InferResultType<TGroqBuilder>>;
-
-export type InferParametersType<TGroqBuilder extends IGroqBuilder<any>> =
-  TGroqBuilder extends IGroqBuilder<any, infer TQueryConfig>
-    ? TQueryConfig["parameters"]
-    : never;
-
-/**
- * Used to store the Result types of a Fragment.
- * This symbol is not used at runtime.
- */
-export declare const FragmentResultTypeTag: unique symbol;
-/**
- * Represents a projection fragment
- */
-export type Fragment<
-  TFragmentResult,
-  _TFragmentInput, // (this type parameter is only included so it shows in an IDE)
-  TProjectionMap
-> = TProjectionMap & {
-  // Track the result type, so we can extract it later:
-  readonly [FragmentResultTypeTag]?: TFragmentResult;
-};
-
-/**
- * Infers the result types of a fragment.
- * @example
- * const productFragment = q.fragment<Product>().project({
- *   name: zod.string(),
- *   price: zod.number(),
- * });
- *
- * type ProductFragment = InferFragmentType<typeof productFragment>;
- */
-export type InferFragmentType<TFragment extends Fragment<any, any, any>> =
-  TFragment extends Fragment<
-    infer TFragmentResult,
-    infer _TFragmentInput,
-    infer _TProjectionMap
-  >
-    ? TFragmentResult
-    : never;
+export {
+  InferResultItem,
+  InferResultType,
+  IGroqBuilderNotChainable,
+  isGroqBuilder,
+  IGroqBuilder,
+  GroqBuilderConfigType,
+  GroqBuilderResultType,
+  InferParametersType,
+} from "../groq-builder";
